refactor: fix clippy errors & remove unused code
