+++ src/substreams_stream.rs
@@ -57,7 +57,7 @@ fn stream_blocks(
     start_block_num: i64,
     stop_block_num: u64,
 ) -> impl Stream<Item = Result<BlockResponse, Error>> {
-    let mut latest_cursor = cursor.unwrap_or_else(|| "".to_string());
+    let mut latest_cursor = cursor.unwrap_or_default();
     let mut backoff = ExponentialBackoff::from_millis(500).max_delay(Duration::from_secs(45));
 
     try_stream! {
